requirements.txt
athentoimaging==0.1
awscli==1.14.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
botocore==1.8.7
certifi==2015.11.20.1
colorama==0.3.7
colour==0.1.2
cv2==1.0
derpconf==0.7.3
docutils==0.14
futures==3.0.4
jmespath==0.9.3
libthumbor==1.3.1
numpy==1.10.4
opencv-engine==1.0.1
pbr==1.8.1
pexif==0.15
Pillow==2.9.0
pyasn1==0.4.2
pycrypto==2.6.1
pycurl==7.19.5.3
pytesseract==0.1.6
python-dateutil==2.6.1
python-magic==0.4.10
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.12
scipy==0.17.0
singledispatch==3.4.0.3
six==1.10.0
statsd==3.2.1
stevedore==1.11.0
tesseract==0.1.3
thumbor==5.2.1
tornado==4.3
virtualenv==14.0.6
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
yolk==0.4.3
